fixed error in do_dTdz_i_proc.m
The function chi_merge_and_avg was changed to accept an averaging
length and now it saves files like dTdz_600sec.mat instead of dTdz.mat.
Added lines of code that now copy the file that exists instead of only
copying dTdz.mat, which doesn’t exist in most cases now.

+++ driver/do_dTdz_i_proc.m
@@ -62,7 +62,11 @@ if do_proc
    chi_merge_and_avg(basedir, 'dTdz', 600);
 
  %_____________________cp result to the input directory______________________
-   ! cp ../proc/dTdz.mat ../input/dTdz_i.mat
+   if exist('../proc/dTdz.mat','file') == 2
+      ! cp ../proc/dTdz.mat ../input/dTdz_i.mat
+   elseif exist('../proc/dTdz_600sec.mat','file') == 2
+      ! cp ../proc/dTdz_600sec.mat ../input/dTdz_i.mat
+   end
 end
 
 %_____________________plotting______________________
